fix: Keep only unique abilities
By some reason some abilities have the same name but different ID. This
behavior cause broken compilation so we have to keep only unique data.

We rely on the following idea: if names are the same, picking the
smallest index is the best option. Unfortunately, UNLOADUNIT is
exception which I have no idea how to handle, except filter out
manually.

Also, restructure the code to make it simpler for tests.

Signed-off-by: Alexander Kurbatov <Alexander.Kurbatov@acronis.com>